feat: Change kebab-case to PascalCase
