+++ src/lib/components/overview-page/ContributorGrid.svelte
